% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{asset.val}
\alias{asset.val}
\title{The asset.val function}
\usage{
asset.val(init.val, pay.schedule)
}
\arguments{
\item{init.val}{is the starting value of the lease, or the amount of money lent}

\item{pay}{schedule is a vector of payments in the lease}
}
\value{
a vector representing the value of the asset over the pay.schedule
}
\description{
This function calculates the asset value of a lease given an initial value and a payment schedule
}
\examples{
asset.val(init.val = 1200, pay.schedule = c(200,0,0,200,0,200,0,200,0,200,200,0,0,200,200,0,200,0,200))
}

