% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{calc.average}
\alias{calc.average}
\title{The calc.average function}
\usage{
calc.average(leases = test.lease.combinations,
  assumptions = default.constants.list, default.rate = norm.default.rate,
  early.default.rate = norm.early.default.rate)
}
\arguments{
\item{leases,}{a data.frame where each row is a lease}

\item{assumptions,}{list of constant values}

\item{default.rate,}{a vector representing the default rate, must be 1000 length with values between 0 and 1}

\item{early.default.rate,}{a vector representing the early default rate, must be 1000 length with values between 0 and 1}
}
\value{
a data frame representing cash flow metrics of the average of input leases
}
\description{
Calculates an average lease given a large distribution of leases
}

