% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{cash.flow}
\alias{cash.flow}
\title{The cash.flow function}
\usage{
cash.flow(hist.leases = test.collect.past,
  average.lease = test.average.lease,
  lease.growth = test.projections.by.day,
  assumptions = default.constants.list, emp.sheet = default.employees)
}
\arguments{
\item{hist.leases,}{data frame with information of historical leases}

\item{average.lease,}{data on cash metrics of the average lease}

\item{lease.growth,}{growth projections for future leases}

\item{assumptions,}{a list of constant values}

\item{emp.sheet,}{information about employees}
}
\value{
a data frame with aggregate cash flow metrics over a period of time
}
\description{
This function calculates cash flow metrics over a period of time given growth projections,
  average lease data, and historical lease data.
}

