% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{cum.default.rate}
\alias{cum.default.rate}
\title{The cum.default.rate function}
\usage{
cum.default.rate(a.lease = test.average.lease, file = NULL,
  text.size = 12, small.text.size = 4,
  assumptions = default.constants.list, default.rate = norm.default.rate)
}
\arguments{
\item{a.lease,}{average lease data}

\item{file,}{if this isn't null, the graph will be saved as this file}

\item{text.size,}{text size for graph titles and sides}

\item{small.text.size,}{text size for the smaller text inside the graph}

\item{assumptions,}{a list of constant values}

\item{default.rate,}{a vector representing the default rate, must be 1000 in length with values between 0 and 1}
}
\value{
The graph of default rate by month
}
\description{
Graphs the cumulative % of payers who have defaulted by month.
}

