% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{debt.calc}
\alias{debt.calc}
\title{The debt.calc function}
\usage{
debt.calc(df = test.cash.flow, assumptions = default.constants.list,
  il.sheet = default.int.lev)
}
\arguments{
\item{df,}{the data frame output from the cash.flow function}

\item{assumptions,}{a list of constant values}

\item{il.sheet,}{information about interest and leverage}
}
\value{
a data frame with debt metrics added on to cash flow metrics
}
\description{
This function calculates debt metrics given cash flow data
}

