% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{incomeStatement}
\alias{incomeStatement}
\title{The incomeStatement function}
\usage{
incomeStatement(file = NULL, mthdata = test.incomeSummary)
}
\arguments{
\item{file,}{the name of the file the sheet will be saved as. If null the file won't be esaved
and the workbook will be returned.}

\item{mthdata,}{cash flow information by year, from the mthSummary function}

\item{yrdata,}{cash flow information by year, from the yrSummary function}

\item{avg.data,}{data representing an average lease, from the calc.average function}

\item{assumptions,}{constants represented in the original data frame form}

\item{d.assumptions,}{constants related to interest and leverage in original data frame form}

\item{emp.data,}{data on employees}

\item{growth.data,}{growth.rates data}

\item{d.rate,}{default rate expressed as length 1000 vector with numbers between 0 and 1}
}
\value{
a data frame with debt metrics added on to cash flow metrics
}
\description{
This function formats results into an income statement sheet, optionally saving it to a file
}

