% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{irr.day}
\alias{irr.day}
\title{The irr.day function}
\usage{
irr.day(df = test.average.lease, no.days = 1000)
}
\arguments{
\item{df,}{average.lease data}

\item{no.days,}{number of days to calculate for}
}
\value{
a number, the IRR
}
\description{
Calculates IRR on a by-day basis
}

