% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{irr.month}
\alias{irr.month}
\title{The irr.month function}
\usage{
irr.month(df = test.average.lease, no.months = 24)
}
\arguments{
\item{df,}{average.lease data}

\item{no.days,}{number of days to calculate for}
}
\value{
a number, the IRR
}
\description{
Calculates IRR on a by-month basis
}

