% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{lease.combinations}
\alias{lease.combinations}
\title{The lease.combinations function}
\usage{
lease.combinations(lease.att = default.lease.att)
}
\arguments{
\item{lease.att,}{a list of lists that represent different distributions of lease traits}
}
\value{
a data frame where each row represents a lease
}
\description{
This function calculates the all possible combinations of leases given a set of attributes
}

