% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{lease.returns}
\alias{lease.returns}
\title{The lease.returns function}
\usage{
lease.returns(a.lease = test.average.lease,
  assumption.list = default.constants.list)
}
\arguments{
\item{a.lease,}{data on cash metrics of the average lease}

\item{assumption.list,}{a list of constant values}
}
\value{
a list with 12 month, 18 month, and total returns
}
\description{
This calculates 12 month, 18 month, and total returns given average lease data
}

