% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{mthIncomeSummary}
\alias{mthIncomeSummary}
\title{The mthIncomeSummary function}
\usage{
mthIncomeSummary(df = test.output.subset)
}
\arguments{
\item{df,}{cash flow data}
}
\value{
a data.frame with data organized by month
}
\description{
Summarizes cash flow data by month
}

