% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{mthSummary}
\alias{mthSummary}
\title{The mthSummary function}
\usage{
mthSummary(df = test.output.subset)
}
\arguments{
\item{df,}{cash flow data}
}
\value{
a data.frame with data organized by month
}
\description{
Summarizes cash flow data by month
}

