% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{output.subset}
\alias{output.subset}
\title{The output.subset function}
\usage{
output.subset(df = test.debt.calc, first.month = default.output.start.month,
  last.month = default.output.end.month)
}
\arguments{
\item{df,}{the data frame to subset}

\item{first.month,}{the first month to be shown}

\item{last.month,}{the last month to be shown}
}
\value{
the subset of df within first.month and last.month
}
\description{
This function subsets a data frame to a desired time frame
}

