% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{prd}
\alias{prd}
\title{The prd function}
\usage{
prd(co.indicator = FALSE, past.payments, past.pay.dates, installments,
  installment.states, installments.balance, installment.dates,
  installment.pay.dates, installment.indexes, pay.gap, start.index, start.date,
  paid.buckets = c(0, 5, 20, 40, 1000), paid.bucket.vals = c(0.1, 0.5, 1, 2,
  4), new.buckets = c(0, 30, 60, 1000), new.bucket.vals = c(0.5, 0.5, 0.75,
  1), assumptions = default.constants.list,
  default.rate = norm.default.rate)
}
\arguments{
\item{co.indicator,}{a boolean indicating whether the lease is charged off.
If charged off, predicted payments vector is just 0s.}

\item{past.payments,}{A vector containing past payments.}

\item{past.pay.dates,}{A vector containing the dates of past payments (corresponds 1:1 with past.payments)}

\item{installments,}{A vector with all the installments of the lease}

\item{installment.states,}{A vector containing the states of all installments (corresponds 1:1 with installments)}

\item{installments.balance,}{A vector with the balance remaining of all installments. Is 0 if the installment is
paid. (corresponds 1:1 with installments)}

\item{installment.dates,}{A vector containing the dates each installment is due (corresponds 1:1 with installments)}

\item{installment.pay.dates,}{A vector containing the dates paid installments were paid. Corresponds 1:1 with installments,
but is NA for installments that have not been paid.}

\item{installment.indexes,}{A vector containing the indexes of each installment in the larger data frame. (corresponds
1:1 with installments)}

\item{pay.gap,}{An integer representing the number of days between scheduled payments}

\item{start.index,}{An integer representing the index matching this lease's start date in the larger data frame}

\item{start.date,}{The start date of the lease}

\item{paid.buckets,}{Vector of integers representing different "buckets" of ranges for leases that have already been partially
paid. Should be same lenghth as paid.bucket.vals}

\item{paid.bucket.vals,}{Vector of values that correspond to each bucket in paid.buckets. Index 1 in paid.bucket.vals
corresponds with values under index 1 in paid.buckets, Index 2 in paid.bucket.vals corresponds with values between
indexes 1 and 2 in paid.buckets, and so on.}

\item{new.buckets,}{Vector of integers representing different "buckets" of ranges for brand new lesases. Should be same length as
new.bucket.vals}

\item{new.bucket.vals:}{Vector of values that correspond to each bucket in new.buckets.Indexes correspond the same way as paid.bucket.vals
and paid.buckets.}

\item{assumptions:}{constant values, represented as a list}

\item{default.rate:}{default rate represented as a vector 1000 long with values between 0 and 1}
}
\value{
A vector representing the predicted payments for the future.
}
\description{
This function calculates a predicted payments vector given data about previous payments and installments
}

