% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{projections.by.day}
\alias{projections.by.day}
\title{The projections.by.day function}
\usage{
projections.by.day(start.date = head(as.Date(default.growth.rates$Dates,
  "\%m/\%d/\%Y"), 1), end.date = tail(as.Date(default.growth.rates$Dates,
  "\%m/\%d/\%Y"), 1), growth.projections = data.frame(date =
  as.Date(as.vector(default.growth.rates$Dates), "\%m/\%d/\%Y"), rate =
  as.vector(default.growth.rates$Leases), per.what = 28))
}
\arguments{
\item{start.date,}{the first date for projections to be shown}

\item{growth.projections,}{a dataframe with dates, rates, and a "per what"}

\item{end}{date, the last date for projections to be shown}
}
\value{
a projections for each day
}
\description{
This function projects future lease volume given a start date, end date, and projections
}

