% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{reverse.irr}
\alias{reverse.irr}
\title{The reverse.irr function}
\usage{
reverse.irr(init.val, pay.schedule)
}
\arguments{
\item{init.val}{is the starting value of the lease, or the amount of money lent}

\item{pay}{schedule is a vector of payments in the lease, make sure it has no gaps in the schedule}
}
\value{
a number representing the IRR of the inputs
}
\description{
This function calculates the irr of a lease (as if it was a loan) given an
initial value and a payment schedule, used here as a helper for asset.val
}
\examples{
reverse.irr(init.val = 1200, pay.schedule = rep(200,10))
}

