% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{yrSummary}
\alias{yrSummary}
\title{The yrSummary function}
\usage{
yrSummary(df = test.output.subset)
}
\arguments{
\item{df,}{cash flow data}
}
\value{
a data.frame with data organized by year
}
\description{
Summarizes cash flow data by year
}

