% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeaseFunctions.R
\name{saveToSheet}
\alias{saveToSheet}
\title{The saveToSheet function}
\usage{
saveToSheet(file = NULL, yrdata = test.yrSummary,
  mthdata = test.mthSummary, avg.data = test.average.lease,
  assumptions = default.constants.table, d.assumptions = default.int.lev,
  emp.data = default.employees, growth.data = default.growth.rates)
}
\arguments{
\item{file,}{the name of the file the sheet will be saved as. If null the file won't be esaved
and the workbook will be returned.}

\item{yrdata,}{cash flow information by year, from the yrSummary function}

\item{mthdata,}{cash flow information by year, from the mthSummary function}

\item{avg.data,}{data representing an average lease, from the calc.average function}

\item{assumptions,}{constants represented in the original data frame form}

\item{d.assumptions,}{constants related to interest and leverage in original data frame form}

\item{emp.data,}{data on employees}

\item{growth.data,}{growth.rates data}
}
\value{
a data frame with debt metrics added on to cash flow metrics
}
\description{
This function writes the results into an excel spreadsheet and saves it
}

